requirements.txt
numpy
google-auth-oauthlib
google-api-python-client
python-dotenv
sqlalchemy
pandas
tqdm